requirement.txt
google-cloud-bigquery==3.3.6
pandas==1.3.3
pendulum==2.1.2
flask==2.2
attrs==22.2
